move inspectors to utils
