+++ core/include/detray/utils/inspectors.hpp
