+++ tests/unit_tests/cpu/material_interaction.cpp
@@ -25,8 +25,8 @@
 #include "detray/propagator/rk_stepper.hpp"
 #include "detray/simulation/random_scatterer.hpp"
 #include "detray/test/types.hpp"
+#include "detray/utils/inspectors.hpp"
 #include "detray/utils/statistics.hpp"
-#include "tests/common/tools/inspectors.hpp"
 
 // VecMem include(s).
 #include <vecmem/memory/host_memory_resource.hpp>
