+++ tests/unit_tests/cpu/tools_navigator.cpp
@@ -12,7 +12,7 @@
 #include "detray/propagator/navigator.hpp"
 #include "detray/test/types.hpp"
 #include "detray/tracks/tracks.hpp"
-#include "tests/common/tools/inspectors.hpp"
+#include "detray/utils/inspectors.hpp"
 
 // VecMem include(s).
 #include <vecmem/memory/host_memory_resource.hpp>
