+++ tutorials/src/cpu/propagation/navigation_inspection.cpp
@@ -15,7 +15,7 @@
 #include "detray/propagator/propagator.hpp"
 #include "detray/simulation/event_generator/track_generators.hpp"
 #include "detray/tracks/tracks.hpp"
-#include "tests/common/tools/inspectors.hpp"
+#include "detray/utils/inspectors.hpp"
 #include "tests/common/tools/particle_gun.hpp"
 
 // Example linear algebra plugin: std::array
